requirements.txt
tkinter
rarfile
tqdm
pikepdf
msoffcrypto-tool
py7zr 